generic entry list item + rename files and vars
