+++ ui/src/components/EntryListItem/KafkaEntryContent.tsx
@@ -1,0 +1,15 @@
+import {Entry} from "./EntryListItem";
+import React from "react";
+
+export interface KafkaEntry extends Entry{
+}
+
+interface KafkaEntryContentProps {
+    entry: KafkaEntry;
+}
+
+export const KafkaEntryContent: React.FC<KafkaEntryContentProps> = ({entry}) => {
+
+    return <>
+    </>
+}
