+++ ui/src/components/UI/FilterSelect.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { MenuItem } from '@material-ui/core';
-import style from './style/HARFilterSelect.module.sass';
+import style from '../style/FilterSelect.module.sass';
 import { Select, SelectProps } from "./Select";
 
 interface HARFilterSelectProps extends SelectProps {
@@ -12,7 +12,7 @@ interface HARFilterSelectProps extends SelectProps {
     transformDisplay?: (string) => string;
 }
 
-export const HARFilterSelect: React.FC<HARFilterSelectProps> = ({items, value, onChange, label, allowMultiple= false, transformDisplay}) => {
+export const FilterSelect: React.FC<HARFilterSelectProps> = ({items, value, onChange, label, allowMultiple= false, transformDisplay}) => {
     return <Select
         value={value}
         multiple={allowMultiple}
