+++ ui/src/components/UI/Select.tsx
@@ -1,8 +1,8 @@
-import {ReactComponent as DefaultIconDown} from './assets/default_icon_down.svg';
+import {ReactComponent as DefaultIconDown} from '../assets/default_icon_down.svg';
 import {MenuItem, Select as MUISelect} from '@material-ui/core';
 import React from 'react';
 import {SelectProps as MUISelectProps} from '@material-ui/core/Select/Select';
-import styles from './style/Select.module.sass';
+import styles from '../style/Select.module.sass';
 
 export const ALL_KEY= 'All';
