+++ ui/src/components/UI/SyntaxHighlighter/highlighterStyle.ts
