+++ ui/src/components/UI/SyntaxHighlighter/index.scss
