+++ ui/src/components/style/EntriesList.module.sass
