+++ ui/src/components/style/EntryDetailed.module.sass
