--- ui/src/components/style/HarPaging.module.sass
@@ -1,16 +1,0 @@
-.HarPaging
-  justify-content: center
-  align-items: center
-  padding-bottom: 10px
-
-  img
-    cursor: pointer
-
-.text
-  color: #8f9bb2
-  font-size: 14px
-  padding: 0 10px
-
-.pageNumber
-  color: #fff
-  font-weight: 600
