+++ ui/src/helpers/api.js
@@ -3,16 +3,16 @@ import * as axios from "axios";
 const mizuAPIPathPrefix = "/mizu";
 
 // When working locally (with npm run start) change to:
-// export const MizuWebsocketURL = `ws://localhost:8899/ws`;
-export const MizuWebsocketURL = `ws://${window.location.host}${mizuAPIPathPrefix}/ws`;
+export const MizuWebsocketURL = `ws://localhost:8899/ws`;
+// export const MizuWebsocketURL = `ws://${window.location.host}${mizuAPIPathPrefix}/ws`;
 
 export default class Api {
 
     constructor() {
 
         // When working locally (with npm run start) change to:
-        // const apiURL = `http://localhost:8899/api/`;
-        const apiURL = `${window.location.origin}${mizuAPIPathPrefix}/api/`;
+        const apiURL = `http://localhost:8899/api/`;
+        // const apiURL = `${window.location.origin}${mizuAPIPathPrefix}/api/`;
 
         this.client = axios.create({
             baseURL: apiURL,
